Fix build issues on linux

+++ fs/serve_linux_test.go
@@ -15,9 +15,9 @@ func platformStatfs(st *syscall.Statfs_t) *statfsResult {
 		Bavail:  st.Bavail,
 		Files:   st.Files,
 		Ffree:   st.Ffree,
-		Bsize:   st.Bsize,
-		Namelen: st.Namelen,
-		Frsize:  st.Frsize,
+		Bsize:   int64(st.Bsize),
+		Namelen: int64(st.Namelen),
+		Frsize:  int64(st.Frsize),
 	}
 }
 
@@ -27,10 +27,10 @@ func platformStat(fi os.FileInfo) *statResult {
 	}
 	st := fi.Sys().(*syscall.Stat_t)
 	r.Ino = st.Ino
-	r.Nlink = st.Nlink
+	r.Nlink = uint64(st.Nlink)
 	r.UID = st.Uid
 	r.GID = st.Gid
-	r.Blksize = st.Blksize
+	r.Blksize = int64(st.Blksize)
 	return r
 }
